downloadmoreRAM.com
Force store all printed strings to FLASH on compile - reduced RAM
consumption from 45% -> 25%!

+++ SeaSense_ArduinoLib/globals.h
@@ -21,7 +21,7 @@
         extern double Temp;
         extern unsigned int Depth;
         extern int Cond;
-        extern unsigned int Light;
+        extern unsigned long Light;
         extern int carryOut;
         extern int Head;
         extern int AccelX,AccelY,AccelZ;
